rename EthWallet feature to EvmWallet
