+++ README.md
@@ -34,7 +34,7 @@ the Astria bridge.
 * `src/features` - Organizes components, contexts, hooks, services, types, and
   utils for different features
   * `src/features/CosmosWallet` - User for interacting with Keplr wallet
-  * `src/features/EthWallet` - Used for interacting with EVM wallets
+  * `src/features/EvmWallet` - Used for interacting with EVM wallets
   * `src/features/Notifications` - Used for displaying notifications and toasts
 * `src/pages`
   * React components for each page
