+++ web/src/features/EvmWallet/utils/utils.test.ts
